+++ database/read_test.go
@@ -398,37 +398,37 @@ func TestDatabaseCount(t *testing.T) {
 	setupTestData(t)
 
 	// Test basic count - should return total number of records
-	count := new(int64)
+	count := new(int)
 	require.NoError(t, database.Database[*TestUser](context.Background()).Count(count))
-	require.Equal(t, int64(3), *count, "should have 3 records")
+	require.Equal(t, 3, *count, "should have 3 records")
 
 	// Test count with query conditions
 	require.NoError(t, database.Database[*TestUser](context.Background()).WithQuery(&TestUser{Name: u1.Name}).Count(count))
-	require.Equal(t, int64(1), *count, "should have 1 record matching name")
+	require.Equal(t, 1, *count, "should have 1 record matching name")
 
 	require.NoError(t, database.Database[*TestUser](context.Background()).WithQuery(&TestUser{Age: u2.Age}).Count(count))
-	require.Equal(t, int64(1), *count, "should have 1 record matching age")
+	require.Equal(t, 1, *count, "should have 1 record matching age")
 
 	// Test count after soft delete - soft-deleted records should not be counted
 	require.NoError(t, database.Database[*TestUser](context.Background()).Delete(u1))
 	require.NoError(t, database.Database[*TestUser](context.Background()).Count(count))
-	require.Equal(t, int64(2), *count, "should have 2 records after soft delete")
+	require.Equal(t, 2, *count, "should have 2 records after soft delete")
 
 	// Test count with query after soft delete
 	require.NoError(t, database.Database[*TestUser](context.Background()).WithQuery(&TestUser{Name: u1.Name}).Count(count))
-	require.Equal(t, int64(0), *count, "soft-deleted record should not be counted")
+	require.Equal(t, 0, *count, "soft-deleted record should not be counted")
 
 	// Test count multiple times - should be idempotent
 	require.NoError(t, database.Database[*TestUser](context.Background()).Count(count))
-	require.Equal(t, int64(2), *count)
+	require.Equal(t, 2, *count)
 	require.NoError(t, database.Database[*TestUser](context.Background()).Count(count))
-	require.Equal(t, int64(2), *count)
+	require.Equal(t, 2, *count)
 
 	// Test count with different model types
 	require.NoError(t, database.Database[*TestProduct](context.Background()).Count(count))
-	require.GreaterOrEqual(t, *count, int64(0), "product count should be non-negative")
+	require.GreaterOrEqual(t, *count, 0, "product count should be non-negative")
 	require.NoError(t, database.Database[*TestCategory](context.Background()).Count(count))
-	require.GreaterOrEqual(t, *count, int64(0), "category count should be non-negative")
+	require.GreaterOrEqual(t, *count, 0, "category count should be non-negative")
 
 	// Test count with nil pointer - should return error
 	err := database.Database[*TestUser](context.Background()).Count(nil)
